add check if theme exists

+++ customizations/scripts-common.sh
@@ -6,23 +6,27 @@ case "$response" in
   [yY][eE][sS]|[yY])
     read -e -p "==> ksa: What would you like to call the new theme? (kage): " themename
     [ -z "${themename}" ] && themename='kage'
-    echo '==> ksa: Cloning kage starter theme into themes directory and removing its .git directory'
-    git clone git@github.com:kehittamo/kage.git htdocs/wp-content/themes/$themename
-    rm -rf htdocs/wp-content/themes/$themename/.git
-    if [ -f gulp.config.js.example ] && [ ! -f gulp.config.js ]; then
-      cp gulp.config.js.example gulp.config.js
-      sed -i '' -e "s/kage/$themename/g" gulp.config.js
+    if [ ! -d htdocs/wp-content/themes/$themename ]; then
+      echo '==> ksa: Cloning kage starter theme into themes directory and removing its .git directory'
+      git clone git@github.com:kehittamo/kage.git htdocs/wp-content/themes/$themename
+      rm -rf htdocs/wp-content/themes/$themename/.git
+      if [ -f gulp.config.js.example ] && [ ! -f gulp.config.js ]; then
+        cp gulp.config.js.example gulp.config.js
+        sed -i '' -e "s/kage/$themename/g" gulp.config.js
+      fi
+      if [ ! $themename = "kage" ]; then
+        echo "==> ksa: Setting up theme $themename"
+        mv htdocs/wp-content/themes/$themename/lang/kage.pot htdocs/wp-content/themes/$themename/lang/$themename.pot
+        sed -i '' -e "s/Kage/$themename/g" htdocs/wp-content/themes/$themename/style.css
+        vagrant ssh -- -t "wp theme activate $themename"
+      fi
+      cd htdocs/wp-content/themes/$themename
+      bower install
+      cd ../../../../
+      echo "==> ksa: Theme $themename succesfully initialized and activated."
+    else
+      echo "==> ksa: Theme $themename directory already exists. Theme not initialized."
     fi
-    if [ ! $themename = "kage" ]; then
-      echo "==> ksa: Setting up theme $themename"
-      mv htdocs/wp-content/themes/$themename/lang/kage.pot htdocs/wp-content/themes/$themename/lang/$themename.pot
-      sed -i '' -e "s/Kage/$themename/g" htdocs/wp-content/themes/$themename/style.css
-      vagrant ssh -- -t "wp theme activate $themename"
-    fi
-    cd htdocs/wp-content/themes/$themename
-    bower install
-    cd ../../../../
-    echo "==> ksa: Theme $themename succesfully initialized and activated."
     ;;
   *)
     echo '==> ksa: Theme not initialized.'
